// src/types/ui.ts
export enum ViewState {
    LANDING = 'LANDING',
    ACADEMY_SELECTION = 'ACADEMY_SELECTION',
    DASHBOARD = 'DASHBOARD',
    COURSE_PLAYER = 'COURSE_PLAYER',
    CATALOG = 'CATALOG',
    ADMIN_PANEL = 'ADMIN_PANEL',
    CAREER_CENTER = 'CAREER_CENTER',
    CERTIFICATES = 'CERTIFICATES',
    USER_MANAGEMENT = 'USER_MANAGEMENT',
    REPORTS = 'REPORTS',
    SETTINGS = 'SETTINGS',
    SUPPORT = 'SUPPORT',
    LOGIN = 'LOGIN',
    STUDENT_ANALYTICS = 'STUDENT_ANALYTICS',
    SYSTEM_LOGS = 'SYSTEM_LOGS',
    EDUCATION_EBOOKS = 'EDUCATION_EBOOKS',
    EDUCATION_VIDEOS = 'EDUCATION_VIDEOS',
    EDUCATION_LIVE = 'EDUCATION_LIVE',
    EDUCATION_ASSIGNMENTS = 'EDUCATION_ASSIGNMENTS',
    EDUCATION_QUIZZES = 'EDUCATION_QUIZZES',
    EDUCATION_EXAMS = 'EDUCATION_EXAMS'
}